refactor(data): split ReligionData into partial classes by concern
Decompose the 965-line ReligionData into partials by responsibility
(members, roles, prestige, blessings, bans, activity log, chronicle) plus
a core file holding identity, constructors, and the shared lock. Purely
organizational: ProtoMember field numbers, the single shared lock, and the
public surface are all unchanged, so persistence and all call sites are
unaffected.
